Fix after review. Benchmarks.

+++ storage/txcache/txMeasures.go
@@ -22,8 +22,8 @@ func estimateTxGas(tx data.TransactionHandler) int64 {
 func estimateTxFee(tx data.TransactionHandler) int64 {
 	// In order to obtain the result as micro ERD, we have to divide by 10^12 (since 1 ERD = 10^18 gas currency units)
 	// In order to have better precision, we divide each of the factors by 10^6
-	gasLimit := float64(tx.GetGasLimit()) / 1000000
-	gasPrice := float64(tx.GetGasPrice()) / 1000000
+	gasLimit := float32(tx.GetGasLimit()) / 1000000
+	gasPrice := float32(tx.GetGasPrice()) / 1000000
 	feeInMicroERD := gasLimit * gasPrice
 	return int64(feeInMicroERD)
 }
